% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorF.R
\name{errorF}
\alias{errorF}
\title{errorF}
\usage{
errorF(y, vec)
}
\arguments{
\item{y}{representa la serie de observaciones reales.}

\item{vec}{representa la matriz de observaciones de contraste. Cada columna representa una variable.}
}
\value{
Esta funcion retorna una matriz de la raiz del error cuadrático medio para vectores de contraste ingresados en vec.
}
\description{
Esta funcion se encarga del cálculo de la raíz cuadrada del error cuadrático medio.
La función facilita la comparación entre varios errores de un mismo procedimiento.
}
\examples{
# Example for two-dimensions
x <- c(1, 6)
v <- matrix(c(5, 3,
              1, 4), ncol=2, byrow=TRUE)
errorF(y=x, vec=v)

# Example for three-dimensions
x <- c(1, 6, 5)
v <- matrix(c(5, 3, 1,
              1, 4, 3,
              0, 2, 2), ncol=3, byrow=TRUE)
errorF(y=x, vec=v)

}
\author{
Omar Rios Saavedra, \email{orioss@unal.edu.co}
}
