% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multi}
\alias{multi}
\title{multi}
\usage{
multi(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
Esta funcion retorna un numero que corresponde a la sum of \code{x} and \code{y}.
}
\description{
Esta funcion sirve para obtener la multiplicacion de dos numeros reales.
}
\details{
Esta funcion sirva para bla bla bla
en este parrafo Omar debe dar todos los detalles tecnicos de la funcion
aqui es donde explica lo que considere que se necesita.
}
\examples{
multi(x=2, y=5)
multi(4, y=5)
multi(x=7, 3)
multi(2, 2)
}
\author{
Valentina Hurtado Sepúlveda, \email{vhurtados@unal.edu.co}
}
